Add Product Relation Checks Before Deleting Categories & SubCategories

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Backend;
 
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\Product;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
@@ -100,11 +101,19 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category = Category::findOrFail($id);
+
         $subCategory = SubCategory::where('category_id', $category->id)->count();
+
         if ($subCategory > 0) {
             return response(['status' => 'error', 'message' => 'This Items Contain, Sub Items For Delete This you Have to Delete the Sub Item First!']);
         }
+
+        if (Product::where(['category_id' => $category->id])->count() > 0) {
+            return response(['status' => 'error', 'message' => 'This Item Content Relation Some Products. You cant Delete It.']);
+        }
+
         $category->delete();
+
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
     }
